requirements.txt
uvicorn==0.23.2
fastapi==0.108.0
pydantic==2.2.1
elasticsearch==8.14.0
openai==1.37.1
python-dotenv==1.0.1